requirements.txt
astroid==2.2.5
autopep8==1.4.4
Babel==2.7.0
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.1
defusedxml==0.6.0
Django==2.2.4
django-allauth==0.39.1
django-debug-toolbar==1.10.1
django-phonenumber-field==3.0.1
django-tinymce4-lite==1.7.5
idna==2.8
isort==4.3.20
jsmin==2.2.2
lazy-object-proxy==1.4.1
mccabe==0.6.1
mysqlclient==1.4.2.post1
oauthlib==3.0.1
Pillow==6.1.0
pycodestyle==2.5.0
pylint==2.3.1
pylint-django==2.0.9
pylint-plugin-utils==0.5
python-decouple==3.1
python3-openid==3.1.0
pytz==2018.5
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.2.4
typed-ast==1.4.0
urllib3==1.25.3
wrapt==1.11.2
